+++ transformer_engine/new/common/compute_pipeline.py
@@ -1,6 +1,6 @@
 from typing import Generator, final
 
-from transformer_engine.pytorch.sequential_new.common_back.generic_tensor import (
+from .generic_tensor import (
     GenericTensor,
 )
