+++ transformer_engine/new/common/fusions.py
