+++ transformer_engine/new/pytorch/implementation/compute_pipeline_function.py
@@ -1,6 +1,6 @@
 import torch.autograd
 
-from ...common_back import ComputePipeline
+from ...common import ComputePipeline
 
 
 def apply(x: torch.Tensor, pipeline: ComputePipeline) -> torch.Tensor:
