+++ transformer_engine/new/pytorch/interface/__init__.py
@@ -1,0 +1,12 @@
+from .atomic_modules import Linear, LayerNorm
+from .complex_modules import LayerNormLinear, LayerNormMLP  # type: ignore
+from .meta_modules import Sequential, Residual
+
+__all__ = [
+    "LayerNorm",
+    "LayerNormLinear",
+    "LayerNormMLP",
+    "Linear",
+    "Residual",
+    "Sequential",
+]
