+++ transformer_engine/new/pytorch/interface/meta_modules/__init__.py
@@ -1,0 +1,7 @@
+from .sequential import Sequential
+from .residual import Residual
+
+__all__ = [
+    "Sequential",
+    "Residual",
+]
